local.requirements.txt
absl-py==2.0.0
accelerate==0.27.2
ale-py==0.7.4
alembic==1.13.1
annotated-types==0.6.0
ansiwrap==0.8.4
anyio==4.0.0
apturl==0.5.2
arch==5.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciichart==0.1
asciichartpy==1.5.25
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==19.2.0
autopep8==2.0.4
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
Babel==2.13.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
black==23.10.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boolean.py==4.0
bottle==0.12.25
Brlapi==0.8.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2020.6.20
cffi==1.16.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.4
colorlog==6.8.2
comm==0.1.4
command-not-found==0.3
confection==0.1.4
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
dash==2.15.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.3.0
defer==1.0.6
defusedxml==0.7.1
dgl==1.1.3
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.0
Farama-Notifications==0.0.4
fasteners==0.14.1
fastjsonschema==2.18.1
filelock==3.12.4
flake8==6.1.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
fsspec==2024.2.0
future==0.18.2
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.59.2
gym==0.21.0
gymnasium==0.29.1
h5py==3.10.0
httplib2==0.20.2
huggingface-hub==0.20.3
identify==2.5.30
idna==3.3
importlib-metadata==4.6.4
importlib-resources==6.1.1
iniconfig==2.0.0
intel-openmp==2024.0.2
ipykernel==6.6.1
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.9.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.24.0
jupytext==1.6.0
keras==2.15.0
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
markdown-it-py==0.5.8
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
mkl==2024.0.0
mkl-service==2.4.0
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==2.8.6
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.46
opencv-python==4.8.1.78
opt-einsum==3.3.0
optuna==3.5.0
optuna-dashboard==0.14.0
ordered-set==4.0.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
papermill==2.3.3
paramiko==2.9.3
parso==0.8.3
pathspec==0.11.2
patsy==0.5.6
pep8-naming==0.13.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.0.0
pluggy==1.3.0
pre-commit==3.6.0
preshed==3.0.9
progressbar2==3.51.4
prometheus-client==0.17.1
prompt-toolkit==3.0.39
property-cached==1.6.4
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.11.0
pycparser==2.21
pycups==2.0.1
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.1.0
pyglet==2.0.10
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
python-utils==3.8.1
pytz==2023.3.post1
pyvis==0.3.2
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.4.4
QtPy==2.4.1
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rliable==1.0.8
rsa==4.9
safetensors==0.4.2
sb3-contrib==2.2.1
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.1
SecretStorage==3.3.1
Send2Trash==1.8.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
Shimmy==1.3.0
simpy==4.1.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.4
SQLAlchemy==2.0.27
srsly==2.4.8
stable-baselines==2.10.2
stable-baselines3==2.2.1
stack-data==0.6.3
statsmodels==0.14.1
systemd-python==234
tabulate==0.8.10
tbb==2021.11.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.17.1
textwrap3==0.9.2
thinc==8.2.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==1.13.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.6
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
